:hammer: Add counter task sidebar
